+++ skforecast/__init__.py
@@ -1,2 +1,2 @@
 name = "skforecast"
-__version__ = "0.5.0"
+__version__ = "0.5.1"